refactor(rpc): Move verifier set_ functions outside of rpc module.
